In folder repositories, create a new class MultiSpectrumQueryBuilder with
constructor MultiSpectrumQueryBuilder(List<Spectrum> querySpectra)

method build() that returns a string with SQL query

In SpectrumRepositoryImpl.java, add method multiSpectrumSearch(List<Spectrum> querySpectra) and returns List<SpectrumMatch>

Use the new method multiSpectrumSearch in StudySearchServiceImpl.studySearch() instead of spectrumRepository.spectrumSearch

Compare performances of the new and old methods

+++ src/main/java/org/dulab/adapcompounddb/site/repositories/SpectrumRepositoryImpl.java
@@ -261,7 +261,7 @@ public class SpectrumRepositoryImpl implements SpectrumRepositoryCustom {
         for (final Object[] objects : resultList) {
             final int querySpectrumIndex = ((BigInteger) objects[0]).intValue();
             final long matchSpectrumId = ((BigInteger) objects[1]).longValue();
-            final double score = ((Double) objects[2]).longValue();
+            final double score = ((Double) objects[2]).doubleValue();
 
             final SpectrumMatch match = new SpectrumMatch();
